Refactor appointmetn grouping functionality
