+++ src/Components/Shared/Appointments/List.js
@@ -5,7 +5,10 @@ import { useTranslation } from "react-i18next";
 import useAsyncWithParams from "../../../Hooks/useAsyncWithParams";
 import PropTypes from "prop-types";
 import SharedAPI from "../../../API/SharedAPI";
-import { DateTime } from "luxon";
+import { Collapse, Typography, Grid } from "@material-ui/core";
+import ExpansionToggle from "../../ExpansionToggle";
+import useToggle from "../../../Hooks/useToggle";
+import groupAppointments from "../../../Utility/group-appointments";
 
 export default function AppointmentList({ patientId }) {
   const { t } = useTranslation("translation");
@@ -17,43 +20,36 @@ export default function AppointmentList({ patientId }) {
     initialData: [],
   });
 
-  let grouped = {};
-
-  for (let apt of appointments.sort((a, b) => {
-    return DateTime.fromISO(b.datetime).toMillis() - DateTime.fromISO(a.datetime).toMillis()
-  })) {
-    if (DateTime.fromISO(apt.datetime).diffNow("days").minutes > 0) {
-      if (!grouped.future) {
-        grouped.future = [];
-      } else {
-        grouped.future.push(apt);
-      }
-    } else {
-      if (!grouped.past) {
-        grouped.past = [];
-      } else {
-        grouped.past.push(apt);
-      }
-    }
-  }
+  const grouped = groupAppointments(appointments);
 
   return (
     <>
       {Object.keys(grouped).length > 0 &&
-        Object.keys(grouped).map((each) => {
-          return (
-            <>
-              <p>{each}</p>
-              <Items items={grouped[each]} />
-            </>
-          );
-        })}
+        Object.keys(grouped).map((each) => (
+          <Group key={`group-${each}`} title={each} items={grouped[each]} />
+        ))}
     </>
   );
 }
 
+const Group = ({ items, title }) => {
+  const [expanded, toggle] = useToggle(true);
+  return (
+    <>
+      <Grid onClick={toggle} container wrap="nowrap">
+        <Typography style={{ fontSize: "1.2rem" }} variant="h2">
+          {title} {items.length}
+        </Typography>
+        <ExpansionToggle expanded={expanded} />
+      </Grid>
+      <Collapse in={expanded}>
+        <Items items={items} />
+      </Collapse>
+    </>
+  );
+};
+
 const Items = ({ items }) => {
-  console.log(items);
   return (
     <>
       {items.map((each) => (
